Removed test Code cleanup Fixed problem with flag subcommand
